Put cisco_prime agent in separate file
CMK-8881
