+++ cmk/gui/plugins/wato/datasource_programs.py
@@ -335,88 +335,6 @@ rulespec_registry.register(
 )
 
 
-def _valuespec_special_agents_cisco_prime():
-    return Dictionary(
-        elements=[
-            (
-                "host",
-                CascadingDropdown(
-                    choices=[
-                        ("ip_address", _("IP Address")),
-                        ("host_name", _("Host name")),
-                        (
-                            "custom",
-                            _("Custom Host"),
-                            Dictionary(
-                                elements=[
-                                    (
-                                        "host",
-                                        TextInput(
-                                            title=_("Custom Host"),
-                                            allow_empty=False,
-                                        ),
-                                    )
-                                ],
-                                optional_keys=[],
-                            ),
-                        ),
-                    ],
-                    default_value="ip_address",
-                    title=_("Host to use for connecting to Cisco Prime"),
-                ),
-            ),
-            (
-                "basicauth",
-                Tuple(
-                    title=_("BasicAuth settings (optional)"),
-                    help=_("The credentials for api calls with authentication."),
-                    elements=[
-                        TextInput(title=_("Username"), allow_empty=False),
-                        PasswordFromStore(title=_("Password of the user"), allow_empty=False),
-                    ],
-                ),
-            ),
-            ("port", Integer(title=_("Port"), default_value=8080)),
-            (
-                "no-tls",
-                FixedValue(
-                    value=True,
-                    title=_("Don't use TLS/SSL/Https (unsecure)"),
-                    totext=_("TLS/SSL/Https disabled"),
-                ),
-            ),
-            (
-                "no-cert-check",
-                FixedValue(
-                    value=True,
-                    title=_("Disable SSL certificate validation"),
-                    totext=_("SSL certificate validation is disabled"),
-                ),
-            ),
-            (
-                "timeout",
-                Integer(
-                    title=_("Connect Timeout"),
-                    help=_("The network timeout in seconds"),
-                    default_value=60,
-                    minvalue=1,
-                    unit=_("seconds"),
-                ),
-            ),
-        ],
-        title=_("Cisco Prime"),
-    )
-
-
-rulespec_registry.register(
-    HostRulespec(
-        group=RulespecGroupDatasourceProgramsOS,
-        name="special_agents:cisco_prime",
-        valuespec=_valuespec_special_agents_cisco_prime,
-    )
-)
-
-
 def _transform_kubernetes_connection_params(value):
     """Check_mk version 2.0: rework input of connection paramters to improve intuitive use.
     Note that keys are removed from the parameters dictionary!
